u-boot-artik53x.bbappend: Add patch to load the RAPTOR dtb by default
It was noticed on the Neurosity board that, sometimes, it loads the
COMPY variant dtb instead of the RAPTOR variant. The issue reported
was that i2c devices i2c-8 and i2c-9 were now present in /dev.
A patch over u-boot is added to remove the checkup and loading of the
COMPY variant and always load the RAPTOR variant of the dtb.

Changelog-entry: Force the loading the RAPTOR board variant dtb
Signed-off-by: Vicentiu Galanopulo <vicentiu@balena.io>